0202 ham master 갈ㄹ라고

+++ classic_shop/src/com/classic/order/daoImp/OrderDaoImp.java
@@ -164,7 +164,7 @@ public class OrderDaoImp implements OrderDAO{
 			coupon.setNum(rs.getInt("coupon_num"));
 			coupon.setMem_num(rs.getInt("mem_num"));
 			coupon.setCoupon_name(rs.getString("coupon_name"));
-			coupon.setSale(rs.getInt("discount"));
+			coupon.setSale(rs.getDouble("discount"));
 			couponList.add(coupon);
 		}
 		return couponList;
